NDV_Code_By_AnushaR_SalaryCalculator
This Python program helps calculate the final salary of an employee by considering the basic salary, a bonus, and a deduction. It begins by taking the basic salary and the bonus percentage from the user, calculates the bonus amount, and then asks for the deduction percentage to compute the deduction amount. Finally, it adds the bonus to the basic salary and subtracts the deduction to get the final salary. The program neatly displays the basic salary, bonus, deduction, and the final salary, making it easy to understand the overall calculation.