NDV_Code_By_AnushaR_CatalogWebsite
This product catalog website showcases a variety of fashion items including dresses, sarees, handbags, heels, and traditional earrings. The website is designed using HTML and CSS, and is fully responsive for both desktop and mobile screens. It includes sections for navigation, product display, and a footer with social media links. Each product is displayed with an image, name, price, and description to give users a clear idea of the item. The layout is clean, user-friendly, and visually appealing.